+++ MySuperApp/Class/Databsae/other/EADataModel/LBaseModel.h
@@ -26,7 +26,7 @@
 @property (nonatomic, assign) BOOL success; /*可以通过这个判断是否请求成功并且有数据*/
 
 @property (retain, nonatomic) NSString *errorMessage;
-@property (nonatomic, assign) int requestTag; /*这个与相应的请求的tag是同步的*/
+@property (nonatomic, assign) NSInteger requestTag; /*这个与相应的请求的tag是同步的*/
 
 
 + (LBaseModel *)modelObjectWithDictionary:(NSDictionary *)dict;
